% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{all_data}
\alias{all_data}
\title{VVIQ and TAS data from various studies}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1478 rows and 16 columns.
}
\source{
Data collected by various teams in online experiments:
\itemize{
\item Ale & Burns (2024): https://doi.org/10.31234/osf.io/kj5d3
(study name "burns")
\item Monzel et al. (2024): https://doi.org/10.1016/j.bionps.2024.100106
(study name "monzel")
\item Ruby (2025): Original data collected by Perrine Ruby at the Centre de
Recherche en Neurosciences de Lyon (study name "ruby")
\item Mas (2025): Original data collected by Marine Mas et al. at Louvain
(study name "mas")
\item Kvamme et al. (2025): https://doi.org/10.31234/osf.io/6fhj4_v1
(study name "kvamme")
}
}
\usage{
all_data
}
\description{
This dataset contains the responses to the Vividness of Visual Imagery
Questionnaire (VVIQ) and the Toronto Alexithymia Scale (TAS) collected
across various studies. It
contains the following:
\itemize{
\item \code{study}: the study from which the data originates (see source below)
\item \code{lang}: language of the study (English or French)
\item \code{id}: unique identifier for each participant
\item \code{sex}
\item \code{gender} (if distinct from sex)
\item \code{age}
\item \code{vviq}: total score on the VVIQ
\item \code{tas}: total score on the TAS
\item \code{tas_identify}: Difficulty Identifying Feelings (DIF) subscale
\item \code{tas_describe}: Difficulty Describing Feelings (DDF) subscale
\item \code{tas_external}: Externally Oriented Thinking (EOT) subscale
\item \code{tas_group}: categorical variable indicating alexithymia status
(alexithymic or non-alexithymic)
\item \code{vviq_group_4}: categorical variable indicating VVIQ group with 4 levels
(aphantasia, hypophantasia, typical, hyperphantasia)
\item \code{vviq_group_3}: categorical variable indicating VVIQ group with 3 levels
(aphantasia, hypophantasia, typical/hyperphantasia)
\item \code{vviq_group_2}: categorical variable indicating VVIQ group with 2 levels
(aphantasia/hypophantasia, typical/hyperphantasia)
\item \code{items}: nested list-column with individual VVIQ and TAS item responses
when available
\item \code{other_data}: nested list-column with other questionnaire data collected
in the studies that are not used in the current analyses
}
}
\keyword{datasets}
