% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayesian_results.R
\name{check_slope_evidence}
\alias{check_slope_evidence}
\title{Check VVIQ slope data for Bayesian GAMs}
\usage{
check_slope_evidence(slopes, digits = 3)
}
\arguments{
\item{slopes}{A \code{\link[modelbased:estimate_slopes]{modelbased::estimate_slopes()}} object containing estimated
slopes.}

\item{digits}{The number of digits to be printed for each value.}
}
\value{
A data frame with slope statistics.
}
\description{
This function checks the visualisation recipe of estimated VVIQ slopes from a
\code{\link[modelbased:estimate_slopes]{modelbased::estimate_slopes()}} object computed on a Bayesian Generalised
Additive Model. The function adds an "Evidence" column to the data,
categorizing slopes as "non null" if the probability of direction (PD)
exceeds 97\% CI, or "uncertain" otherwise. This allows to check manually the
thresholds to choose for the ".f_group" column in the plotting function.
This whole ordeal was necessary because the \code{modelbased} package did not
create the ".groups" column correctly with \code{brms} GAM objects.
}
