% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_aphantasia.R
\name{plot_alexithymia_cutoff}
\alias{plot_alexithymia_cutoff}
\title{Plot Alexithymia cut-off line}
\usage{
plot_alexithymia_cutoff(
  color = "red",
  linetype = "dashed",
  linewidth = 0.3,
  txt_x = 30,
  txt_y = 63,
  txt_size = 1.75,
  label = "Alexithymia cut-off",
  ...
)
}
\arguments{
\item{color}{Color of the line and text. Default is "red".}

\item{linetype}{Type of the line. Default is "dashed".}

\item{linewidth}{Width of the line. Default is 0.3.}

\item{txt_x}{X-coordinate for the text annotation. Default is 30.}

\item{txt_y}{Y-coordinate for the text annotation. Default is 63.}

\item{txt_size}{Size of the text annotation. Default is 1.75.}

\item{label}{Text label for the annotation. Default is "Alexithymia cut-off".}

\item{...}{Additional arguments passed to \code{ggplot2::annotate()}.}
}
\value{
A list containing a horizontal line and text annotation for ggplot2.
}
\description{
This function adds a horizontal line to a ggplot object indicating the
clinical Alexithymia cut-off score, along with a text annotation.
}
