% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{plot_coloured_subjects}
\alias{plot_coloured_subjects}
\title{Plot individual participants coloured by VVIQ group}
\usage{
plot_coloured_subjects(
  x,
  y,
  df = aphantasiaEmotions::all_data,
  alpha = 0.4,
  ...
)
}
\arguments{
\item{x}{Variable for the x-axis.}

\item{y}{Variable for the y-axis.}

\item{df}{Data frame containing the data. Default is \code{all_data}.}

\item{alpha}{Transparency level of the points. Default is 0.4.}

\item{...}{Additional arguments passed to \code{geom_point()}.}
}
\value{
A list containing a \code{geom_point} layer for ggplot2.
}
\description{
This function adds a layer to a ggplot object that plots individual data
points coloured by VVIQ group.
}
