% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bayesian_results.R
\name{plot_gam_slopes}
\alias{plot_gam_slopes}
\title{Plot GAM estimated slopes}
\usage{
plot_gam_slopes(
  slopes,
  .f_groups = 1,
  x_lab = "VVIQ score",
  y_lab = "Slope (score variation per unit change in VVIQ)",
  base_theme = ggplot2::theme_minimal,
  ...
)
}
\arguments{
\item{slopes}{A \code{\link[modelbased:estimate_slopes]{modelbased::estimate_slopes()}} object containing estimated
slopes.}

\item{.f_groups}{A function or expression to create a grouping variable for
categorizing slopes. This is necessary to correctly identify slopes with
non-null evidence when using Bayesian GAM models. The function should return
a grouping variable based on the slope data. Default is \code{1}, which means
all slopes are treated as a single group (it will most likely result in an
error if unchanged).}

\item{x_lab}{Label for the x-axis.}

\item{y_lab}{Label for the y-axis.}

\item{base_theme}{Base ggplot2 theme to use with \code{\link[=theme_pdf]{theme_pdf()}}
(default is \code{ggplot2::theme_minimal}).}

\item{...}{Additional arguments passed to the \code{\link[=theme_pdf]{theme_pdf()}} function for
further customization of the plot theme.}
}
\value{
A ggplot2 object representing the GAM estimated slopes plot.
}
\description{
This function creates a plot of estimated slopes from a Generalized Additive
Model (GAM) using the \code{modelbased} package. It visualizes the
estimated slopes along with confidence intervals and categorizes them based
on whether their 95\% credible intervals exclude zero. Note that in case
Bayesian models are used (such as in this whole project), the \code{modelbased}
package does not create the ".groups" column correctly for GAM objects,
so the user must provide a custom grouping function via the \code{.f_groups}
argument to categorize the slopes appropriately. The
\code{\link[=check_slope_evidence]{check_slope_evidence()}} function can be used to inspect the slope data
and determine suitable thresholds for categorization.
}
\examples{
\dontrun{
gam_tot <- 
  fit_brms_model(
    formula = tas ~ s(vviq), 
    data = all_data,
    prior = c(brms::prior(normal(0, 20), class = "b")),
    file = "models/gam_tot.rds"
  )

slopes_tot <-
  modelbased::estimate_slopes(
  gam_tot, trend = "vviq", by = "vviq", length = 75)

check_slope_evidence(slopes_tot)
plot_gam_slopes(
  slopes_tot,
  .f_groups = dplyr::case_when(
    vviq <= 24 ~ 1,
    vviq <= 45 ~ 2,
    vviq <= 76 ~ 3,
    vviq <= 80 ~ 4
   ),
   y_lab = "TAS variation per unit change in VVIQ",
 )
}
}
