% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{plot_vviq_group_proportions}
\alias{plot_vviq_group_proportions}
\title{Plot VVIQ group proportions of a specified variable across studies}
\usage{
plot_vviq_group_proportions(
  df,
  var,
  bar_alpha = 0.3,
  bar_lw = 0.1,
  prop_txt_size = 1.75,
  marg = 8,
  base_size = 7,
  ...
)
}
\arguments{
\item{df}{Data frame containing the data with the specified variable.}

\item{var}{The variable to be plotted (e.g., vviq_group_4).}

\item{bar_alpha}{Transparency level of the bars.}

\item{bar_lw}{Line width of the bar borders.}

\item{prop_txt_size}{Size of the proportion text labels on the bars.}

\item{marg}{Margin size around the plot.}

\item{base_size}{Base font size for the plot theme.}

\item{...}{Additional arguments passed to the \code{\link[=theme_pdf]{theme_pdf()}} function.}
}
\value{
A ggplot2 object representing the bar plot of VVIQ group proportions
for the specified variable across studies.
}
\description{
This function creates a bar plot showing the proportions of a each VVIQ
sub-group for a given classification (e.g., vviq_group_4, vviq_group_3, etc.)
}
\examples{
plot_vviq_group_proportions(
  all_data, vviq_group_4, 
  base_size = 12, prop_txt_size = 3
)
}
