% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_aphantasia.R
\name{scale_discrete_aphantasia}
\alias{scale_discrete_aphantasia}
\title{Custom discrete scale for Aphantasia groups}
\usage{
scale_discrete_aphantasia(aesthetics = c("color", "fill"), name = NULL, ...)
}
\arguments{
\item{aesthetics}{Aesthetics to apply the scale to. Default is
\code{c("color", "fill")}.}

\item{name}{Name of the scale. Default is \code{NULL}.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_discrete_manual()}.}
}
\value{
A ggplot2 scale object for discrete aesthetics.
}
\description{
This function creates a custom discrete colour scale for Aphantasia groups in
ggplot2 plots.
}
