% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_tools.R
\name{scale_x_aphantasia}
\alias{scale_x_aphantasia}
\title{Custom x-axis scale for Aphantasia groups}
\usage{
scale_x_aphantasia(name = NULL, mult = 0, add = c(0, 0.7), ...)
}
\arguments{
\item{name}{Name of the x-axis.}

\item{mult}{Multiplier for \code{\link[ggplot2:expansion]{ggplot2::expansion()}}.}

\item{add}{Additive for \code{\link[ggplot2:expansion]{ggplot2::expansion()}}.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_x_discrete()}.}
}
\value{
A ggplot2 scale object for the x-axis.
}
\description{
This function creates a custom x-axis scale for Aphantasia groups in ggplot2
plots.
}
