% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_distribution.R
\name{summarise_aph_and_alexi}
\alias{summarise_aph_and_alexi}
\title{Compute counts and proportions of alexithymia groups within VVIQ groups}
\usage{
summarise_aph_and_alexi(df, ...)
}
\arguments{
\item{df}{Data frame containing the data with \code{tas_group} and grouping
variables.}

\item{...}{Grouping variables to summarise the data by (e.g., vviq_group_4).}
}
\value{
A data frame with counts and proportions of alexithymia groups
within the specified grouping variables for each study and for the combined
data.
}
\description{
This function computes the counts and proportions of alexithymia groups
within specified grouping variables (e.g., VVIQ groups) for each study
as well as for the combined data across all studies.
}
\examples{
summarise_aph_and_alexi(all_data, vviq_group_4) |> head()
summarise_aph_and_alexi(all_data, vviq_group_2) |> head()
}
