% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_aphantasia.R
\name{write_sample_info}
\alias{write_sample_info}
\title{Write information about the sample on the plots}
\usage{
write_sample_info(df, n_groups = 2, prefix = "", suffix = "", ...)
}
\arguments{
\item{df}{A dataframe with id and Group columns}

\item{n_groups}{The number of groups to display. Must be either 0, 2 or 4.}

\item{prefix}{A string. Text to add before the number of participants}

\item{suffix}{A string. Text to add after the number of participants}

\item{...}{Other arguments passed to the function. Not used.}

\item{type}{A string indicating the type of information to display. Can be
either "groups" or "clusters".}
}
\value{
A call with the number of participants in each group that can be
used in ggplot2 text elements (such as the title or subtitle).
}
\description{
This function formats the number of participants in the sample or in each
group inside a call object that can be displayed nicely on ggplot2 figures.
It started as a personal fancy to learn how to put text in subscript on
ggplots with plotmath, but it ended up being pretty useful.
}
