% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphantasiaEmotions-package.R
\docType{package}
\name{aphantasiaEmotions-package}
\alias{aphantasiaEmotions}
\alias{aphantasiaEmotions-package}
\title{aphantasiaEmotions: Data Analysis for "The Linear Relationship Between Visual Imagery and Alexithymia Breaks When Imagery Is Absent: Complete Aphantasics Are No More Alexithymic Than Typical Imagers"}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

This project contains the code to analyse VVIQ and TAS data from studies on aphantasia and alexithymia.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://m-delem.github.io/aphantasiaEmotions/}
  \item \url{https://github.com/m-delem/aphantasiaEmotions}
}

}
\author{
\strong{Maintainer}: Maël Delem \email{mael.delem@pm.me} (\href{https://orcid.org/0009-0005-8518-1991}{ORCID})

}
\keyword{internal}
