% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aphantasiaEmotions-package.R
\docType{package}
\name{aphantasiaEmotions-package}
\alias{aphantasiaEmotions}
\alias{aphantasiaEmotions-package}
\title{aphantasiaEmotions: VVIQ and TAS Analyses in Aphantasia and Alexithymia Studies}
\description{
This project contains the code to analyse VVIQ and TAS data from studies on aphantasia and alexithymia.
}
\author{
\strong{Maintainer}: Maël Delem \email{mael.delem@pm.me} (\href{https://orcid.org/0009-0005-8518-1991}{ORCID})

}
\keyword{internal}
