% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data_distribution.R
\name{plot_group_violins}
\alias{plot_group_violins}
\title{Plot group violins with means and confidence intervals}
\usage{
plot_group_violins(
  formula,
  data = all_data,
  dot_size = 0.5,
  box.linewidth = 0.1,
  middle.linewidth = 0.5,
  violin_width = 0.7,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  x_lab = NULL,
  y_lab = "Score",
  breaks = scales::breaks_pretty(10),
  base_theme = ggplot2::theme_minimal,
  axis_relative_size = 1,
  axis_relative_x = 0.8,
  axis_relative_y = 0.85,
  ...
)
}
\arguments{
\item{formula}{A formula specifying the response and grouping variables.}

\item{data}{Data frame containing the data. Default is \code{all_data}.}

\item{dot_size}{Size of the individual data points. Default is 0.5.}

\item{box.linewidth}{Line width of the box in the crossbar. Default is 0.1.}

\item{middle.linewidth}{Line width of the middle line in the crossbar.
Default is 0.5.}

\item{violin_width}{Width of the half-violins. Default is 0.7.}

\item{title}{Title of the plot.}

\item{subtitle}{Subtitle of the plot.}

\item{caption}{Caption of the plot.}

\item{x_lab}{Label for the x-axis.}

\item{y_lab}{Label for the y-axis. Default is "Score".}

\item{breaks}{Breaks for the y-axis. Default is \code{scales::breaks_pretty(10)}.}

\item{base_theme}{Base theme for the plot. Default is
\code{ggplot2::theme_minimal}.}

\item{axis_relative_size}{Relative size of the axis text. Default is 1.}

\item{axis_relative_x}{Relative size of the x-axis text. Default is 1.}

\item{axis_relative_y}{Relative size of the y-axis text. Default is 0.85.}

\item{...}{Additional arguments passed to the \code{\link[=theme_pdf]{theme_pdf()}} function.}
}
\value{
A ggplot object visualising the group data.
}
\description{
This function creates a ggplot object that visualises group data using
half-violins, individual data points, and mean estimates with confidence
intervals.
}
\examples{
plot_group_violins(
 tas ~ vviq_group_4,
 y_lab = "TAS Score",
 base_size = 12
)
}
